orderline endpoints

+++ src/main/java/com/example/tricolv2sb/Service/SupplierService.java
@@ -13,9 +13,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -25,13 +25,20 @@ public class SupplierService implements SupplierServiceInterface {
     private final SupplierMapper supplierMapper;
     private final SupplierRepository supplierRepository;
 
+    @Transactional(readOnly = true)
+    public List<ReadSupplierDTO> fetchAllSuppliers() {
+        List<Supplier> suppliers = supplierRepository.findAll();
+        return suppliers.stream()
+                .map(supplierMapper::toDto)
+                .toList();
+    }
+
     @Transactional
     public ReadSupplierDTO addSupplier(CreateSupplierDTO dto) {
         supplierRepository.findByIce(dto.getIce())
                 .ifPresent(s -> {
                     throw new SupplierAlreadyExistsException(
-                            "Supplier with this ICE: " + dto.getIce() + " already exists"
-                    );
+                            "Supplier with this ICE: " + dto.getIce() + " already exists");
                 });
 
         Supplier supplier = supplierMapper.toEntity(dto);
@@ -39,21 +46,19 @@ public class SupplierService implements SupplierServiceInterface {
         return supplierMapper.toDto(savedSupplier);
     }
 
-
     @Transactional(readOnly = true)
     public Optional<ReadSupplierDTO> fetchSupplier(Long id) {
         return Optional.of(
                 supplierRepository.findById(id)
                         .map(supplierMapper::toDto)
-                        .orElseThrow(() -> new SupplierNotFoundException("Supplier with ID " + id + " not found"))
-        );
+                        .orElseThrow(() -> new SupplierNotFoundException("Supplier with ID " + id + " not found")));
     }
-
 
     @Transactional
     public void deleteSupplier(Long id) {
-        Supplier supplier = supplierRepository.findById(id)
-                .orElseThrow(() -> new SupplierNotFoundException("Supplier with ID " + id + " not found"));
+        if (!supplierRepository.existsById(id)) {
+            throw new SupplierNotFoundException("Supplier with ID " + id + " not found");
+        }
 
         supplierRepository.deleteById(id);
     }
@@ -63,8 +68,7 @@ public class SupplierService implements SupplierServiceInterface {
 
         Supplier existingSupplier = supplierRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND, "Supplier with ID " + id + " not found"
-                ));
+                        HttpStatus.NOT_FOUND, "Supplier with ID " + id + " not found"));
         supplierMapper.updateFromDto(dto, existingSupplier);
         Supplier savedSupplier = supplierRepository.save(existingSupplier);
         return supplierMapper.toDto(savedSupplier);
